Decompose ListItem component
