+++ dashboard/src/components/partials/EventsList/EventsListItem.js
@@ -3,13 +3,24 @@ import PropTypes from "prop-types";
 import gql from "graphql-tag";
 import { withApollo } from "react-apollo";
 
+import Checkbox from "@material-ui/core/Checkbox";
+import IconButton from "@material-ui/core/IconButton";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
+import MoreVert from "@material-ui/icons/MoreVert";
+import RootRef from "@material-ui/core/RootRef";
+import TableCell from "@material-ui/core/TableCell";
 
 import Code from "/components/Code";
 import resolveEvent from "/mutations/resolveEvent";
 import RelativeDate from "/components/RelativeDate";
-import ListItem from "/components/partials/ListItem";
+
+import MenuController from "/components/controller/MenuController";
+
+import ResourceDetails from "/components/partials/ResourceDetails";
+import TableOverflowCell from "/components/partials/TableOverflowCell";
+import TableSelectableRow from "/components/partials/TableSelectableRow";
+
 import NamespaceLink from "/components/util/NamespaceLink";
 import CheckStatusIcon from "/components/CheckStatusIcon";
 
@@ -102,37 +113,57 @@ class EventListItem extends React.Component {
     const { entity, check, timestamp } = event;
 
     return (
-      <ListItem
-        selected={selected}
-        onChangeSelected={onChangeSelected}
-        icon={
-          event.check && <CheckStatusIcon statusCode={event.check.status} />
-        }
-        title={
-          <NamespaceLink
-            namespace={event.namespace}
-            to={`/events/${entity.name}/${check.name}`}
-          >
-            {entity.name} › {check.name}
-          </NamespaceLink>
-        }
-        details={
-          <React.Fragment>
-            Last occurred{" "}
-            <strong>
-              <RelativeDate dateTime={timestamp} />
-            </strong>{" "}
-            and exited with status <strong>{check.status}</strong>.
-            {check.output && (
+      <TableSelectableRow selected={selected}>
+        <TableCell padding="checkbox">
+          <Checkbox
+            color="primary"
+            checked={selected}
+            onChange={e => onChangeSelected(e.target.checked)}
+          />
+        </TableCell>
+        <TableOverflowCell>
+          <ResourceDetails
+            icon={
+              event.check && <CheckStatusIcon statusCode={event.check.status} />
+            }
+            title={
+              <NamespaceLink
+                namespace={event.namespace}
+                to={`/events/${entity.name}/${check.name}`}
+              >
+                {entity.name} › {check.name}
+              </NamespaceLink>
+            }
+            details={
               <React.Fragment>
-                <br />
-                <Code>{check.output}</Code>
+                Last occurred{" "}
+                <strong>
+                  <RelativeDate dateTime={timestamp} />
+                </strong>{" "}
+                and exited with status <strong>{check.status}</strong>.
+                {check.output && (
+                  <React.Fragment>
+                    <br />
+                    <Code>{check.output}</Code>
+                  </React.Fragment>
+                )}
               </React.Fragment>
+            }
+            renderMenu={this.renderMenu}
+          />
+        </TableOverflowCell>
+        <TableCell padding="checkbox">
+          <MenuController renderMenu={this.renderMenu}>
+            {({ open, ref }) => (
+              <RootRef rootRef={ref}>
+                <IconButton onClick={open}>
+                  <MoreVert />
+                </IconButton>
+              </RootRef>
             )}
-          </React.Fragment>
-        }
-        renderMenu={this.renderMenu}
-      />
+          </MenuController>
+        </TableCell>
+      </TableSelectableRow>
     );
   }
 }
