--- dashboard/src/components/partials/ListItem/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ListItem";
