+++ dashboard/src/components/partials/ResourceDetails/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ResourceDetails";
