+++ dashboard/src/components/partials/TableOverflowCell/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TableOverflowCell";
