+++ dashboard/src/components/partials/TableSelectableRow/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TableSelectableRow";
